Reorganise core code for clarity
